+++ lib/src/commands/mp_remove_attr_option_from_element_command.dart
@@ -51,7 +51,7 @@ class MPRemoveAttrOptionFromElementCommand extends MPCommand {
         .executeRemoveAttrOptionFromElement(
           parentMPID: parentMPID,
           attrName: attrName,
-          plaOriginalLineInTH2File: plaOriginalTH2FileLine,
+          newOriginalLineInTH2File: plaOriginalTH2FileLine,
         );
   }